requirements.txt
azure-storage-blob
azure-identity
azure-keyvault-secrets
python-dotenv
